added facade tests class2

+++ trading-system-server/src/main/java/com/wsep202/TradingSystem/domain/exception/CompositeOperatorNullException.java
@@ -1,7 +1,7 @@
 package com.wsep202.TradingSystem.domain.exception;
 
 public class CompositeOperatorNullException extends TradingSystemException{
-    public CompositeOperatorNullException(int id){
+    public CompositeOperatorNullException(long id){
         super("The discount with id: "+id+" activated but it doesn't include operator!");
     }
 }
